requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
sqlparse==0.4.4
tiktoken==0.6.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
